git-svn-id: svn://136.177.114.72/svn_GW/WPhast/trunk@4891 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/vtkPlaneWidget2.cpp
@@ -28,10 +28,8 @@
 #include <vtkTransform.h>
 
 #include <vtkObjectFactory.h> // reqd by vtkStandardNewMacro
-
-//{{
 #include <vtkArrowSource.h>
-//}}
+
 #include "resource.h"
 
 vtkCxxRevisionMacro(vtkPlaneWidget2, "$Revision$");
@@ -71,43 +69,16 @@ vtkPlaneWidget2::vtkPlaneWidget2(void)
 	this->YAxisActor->SetMapper(AxisMapper);
 	this->YAxisActor->GetProperty()->SetColor(0, 1, 0);
 
-	vtkFloatingPointType* pt1 = this->GetPoint1();
+	double* pt1 = this->GetPoint1();
 	for (int i = 0; i < 3; ++i)
 	{
 		this->m_InvisablePosX[i] = pt1[i];
 	}
 
-	// Represent the active plane
-	//
-	this->ActivePlaneSource = vtkPlaneSource::New();
-
-	this->ActivePlaneOutline = vtkPolyData::New();
-
-	vtkPoints *ptsAct = vtkPoints::New();
-	ptsAct->SetNumberOfPoints(4);
-
-	vtkCellArray *outlineAct = vtkCellArray::New();
-	outlineAct->InsertNextCell(4);
-	outlineAct->InsertCellPoint(0);
-	outlineAct->InsertCellPoint(1);
-	outlineAct->InsertCellPoint(2);
-	outlineAct->InsertCellPoint(3);
-
-	this->ActivePlaneOutline->SetPoints(ptsAct);
-	ptsAct->Delete();
-
-	this->ActivePlaneOutline->SetPolys(outlineAct);
-	outlineAct->Delete();
-
-	this->ActivePlaneMapper = vtkPolyDataMapper::New();
-	this->ActivePlaneMapper->SetInput(this->ActivePlaneSource->GetOutput());
-
-	this->ActivePlaneActor = vtkActor::New();
-	this->ActivePlaneActor->SetMapper(this->ActivePlaneMapper);
-	this->ActivePlaneActor->GetProperty()->SetColor(1, 1, 0);
-	this->ActivePlaneActor->GetProperty()->SetOpacity(.7);
-
-	this->PlanePicker->AddPickList(this->ActivePlaneActor);
+	// Add arrow heads to pick list in order to allow
+	// drag/drop/rotate grid
+	this->PlanePicker->AddPickList(this->XAxisActor);
+	this->PlanePicker->AddPickList(this->YAxisActor);
 
 #if defined(_DEBUG)
 	this->m_VisHandleGeometry = vtkSphereSource::New();
@@ -121,37 +92,32 @@ vtkPlaneWidget2::vtkPlaneWidget2(void)
 	this->m_VisHandle->SetMapper(this->m_VisHandleMapper);
 	this->m_VisHandle->GetProperty()->SetColor(1, 0, 1);
 
-	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX);
+	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX[0], this->m_InvisablePosX[1], this->m_InvisablePosX[2]);
 #endif
 
 }
 
 vtkPlaneWidget2::~vtkPlaneWidget2(void)
 {
-	this->XAxisActor->Delete();
-	this->YAxisActor->Delete();
-	this->AxisMapper->Delete();
-	this->AxisSource->Delete();
-
-	this->ActivePlaneSource->Delete();
-	this->ActivePlaneOutline->Delete();
-	this->ActivePlaneMapper->Delete();
-	this->ActivePlaneActor->Delete();
+	if (this->XAxisActor) this->XAxisActor->Delete();
+	if (this->YAxisActor) this->YAxisActor->Delete();
+	if (this->AxisMapper) this->AxisMapper->Delete();
+	if (this->AxisSource) this->AxisSource->Delete();
 
 #if defined(_DEBUG)
-	this->m_VisHandleGeometry->Delete();
-	this->m_VisHandleMapper->Delete();
-	this->m_VisHandle->Delete();
+	if (this->m_VisHandleGeometry) this->m_VisHandleGeometry->Delete();
+	if (this->m_VisHandleMapper)   this->m_VisHandleMapper->Delete();
+	if (this->m_VisHandle)         this->m_VisHandle->Delete();
 #endif
 }
 
 void vtkPlaneWidget2::SetEnabled(int enabling)
 {
-	vtkRenderer *pSaveCurrentRenderer = this->CurrentRenderer;
-	Superclass::SetEnabled(enabling);
-
 	if (enabling)
 	{
+		// called before adding axes
+		Superclass::SetEnabled(enabling);
+
 		// turn off the normal vector
 		this->CurrentRenderer->RemoveActor(this->LineActor);
 		this->CurrentRenderer->RemoveActor(this->ConeActor);
@@ -162,11 +128,11 @@ void vtkPlaneWidget2::SetEnabled(int enabling)
 		this->CurrentRenderer->AddActor(this->XAxisActor);
 		this->CurrentRenderer->AddActor(this->YAxisActor);
 
-		// add active plane
-		this->CurrentRenderer->AddActor(this->ActivePlaneActor);
-
 #ifdef _DEBUG
-		this->CurrentRenderer->AddActor(this->m_VisHandle);
+		if (this->m_VisHandle)
+		{
+			this->CurrentRenderer->AddActor(this->m_VisHandle);
+		}
 #endif
 
 		this->Interactor->Render();
@@ -175,17 +141,27 @@ void vtkPlaneWidget2::SetEnabled(int enabling)
 	else
 	{
 		// remove axes
-		if (pSaveCurrentRenderer)
+		if (this->CurrentRenderer)
 		{
-			pSaveCurrentRenderer->RemoveActor(this->XAxisActor);
-			pSaveCurrentRenderer->RemoveActor(this->YAxisActor);
-			pSaveCurrentRenderer->RemoveActor(this->ActivePlaneActor);
+			ASSERT(this->XAxisActor->GetNumberOfConsumers()       != 0);
+			ASSERT(this->YAxisActor->GetNumberOfConsumers()       != 0);
+
+			this->CurrentRenderer->RemoveActor(this->XAxisActor);
+			this->CurrentRenderer->RemoveActor(this->YAxisActor);
+
+			ASSERT(this->XAxisActor->GetNumberOfConsumers()       == 0);
+			ASSERT(this->YAxisActor->GetNumberOfConsumers()       == 0);
 
 #ifdef _DEBUG
-			pSaveCurrentRenderer->RemoveActor(this->m_VisHandle);
+			ASSERT(this->m_VisHandle->GetNumberOfConsumers()      != 0);
+			this->CurrentRenderer->RemoveActor(this->m_VisHandle);
+			ASSERT(this->m_VisHandle->GetNumberOfConsumers()      == 0);
 #endif
 		}
 		this->Interactor->Render();
+
+		// must be called after removing axes (this->CurrentRender gets freed)
+		Superclass::SetEnabled(enabling);
 	}
 }
 
@@ -258,7 +234,7 @@ void vtkPlaneWidget2::OnCtrlLeftButtonDown(void)
 	if ( path != NULL )
 	{
 		this->State = vtkPlaneWidget2::Spinning;
-		this->HighlightHandle(path->GetFirstNode()->GetProp());
+		this->HighlightHandle(path->GetFirstNode()->GetViewProp());
 	}
 	else
 	{
@@ -325,8 +301,8 @@ void vtkPlaneWidget2::OnMouseMove(void)
 			if ( this->CurrentHandle == this->Handle[0] )
 			{
 				this->Substate = MovingMoveOrigin;
-				vtkFloatingPointType* o = this->GetOrigin();
-				vtkFloatingPointType prevO[3];
+				double* o = this->GetOrigin();
+				double prevO[3];
 				for (int i = 0; i < 3; ++i)
 				{
 					prevO[i] = o[i];
@@ -339,8 +315,8 @@ void vtkPlaneWidget2::OnMouseMove(void)
 			else if ( this->CurrentHandle == this->Handle[1] )
 			{
 				this->Substate = MovingMovePoint1;
-				vtkFloatingPointType* o = this->GetOrigin();
-				vtkFloatingPointType prevO[3];
+				double* o = this->GetOrigin();
+				double prevO[3];
 				for (int i = 0; i < 3; ++i)
 				{
 					prevO[i] = o[i];
@@ -353,8 +329,8 @@ void vtkPlaneWidget2::OnMouseMove(void)
 			else if ( this->CurrentHandle == this->Handle[2] )
 			{
 				this->Substate = MovingMovePoint2;
-				vtkFloatingPointType* o = this->GetOrigin();
-				vtkFloatingPointType prevO[3];
+				double* o = this->GetOrigin();
+				double prevO[3];
 				for (int i = 0; i < 3; ++i)
 				{
 					prevO[i] = o[i];
@@ -382,8 +358,8 @@ void vtkPlaneWidget2::OnMouseMove(void)
 	}
 	else if (this->State == vtkPlaneWidget::Scaling)
 	{
-		vtkFloatingPointType* o = this->GetOrigin();
-		vtkFloatingPointType prevO[3];
+		double* o = this->GetOrigin();
+		double prevO[3];
 		for (int i = 0; i < 3; ++i)
 		{
 			prevO[i] = o[i];
@@ -418,7 +394,7 @@ void vtkPlaneWidget2::OnMouseMove(void)
 	if (this->CurrentRenderer) this->CurrentRenderer->ResetCameraClippingRange();
 
 #if defined(_DEBUG)
-	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX);
+	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX[0], this->m_InvisablePosX[1], this->m_InvisablePosX[2]);
 #endif
 
 	this->Interactor->Render();
@@ -428,12 +404,12 @@ void vtkPlaneWidget2::Spin(void)
 {
 	vtkRenderWindowInteractor *rwi = this->Interactor;
 
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-	vtkFloatingPointType *pt1 = this->PlaneSource->GetPoint1();
-	vtkFloatingPointType *pt2 = this->PlaneSource->GetPoint2();
-	vtkFloatingPointType *center = this->PlaneSource->GetCenter();
+	double *o = this->PlaneSource->GetOrigin();
+	double *pt1 = this->PlaneSource->GetPoint1();
+	double *pt2 = this->PlaneSource->GetPoint2();
+	double *center = this->PlaneSource->GetCenter();
 
-	vtkFloatingPointType disp_obj_center[3];
+	double disp_obj_center[3];
 
 	this->ComputeWorldToDisplay(center[0], center[1], center[2],
 		disp_obj_center);
@@ -446,11 +422,12 @@ void vtkPlaneWidget2::Spin(void)
 		atan2((double)rwi->GetLastEventPosition()[1] - (double)disp_obj_center[1],
 		(double)rwi->GetLastEventPosition()[0] - (double)disp_obj_center[0]);
 
-	newAngle *= vtkMath::RadiansToDegrees();
-	oldAngle *= vtkMath::RadiansToDegrees();
+	newAngle = vtkMath::DegreesFromRadians(newAngle);
+	oldAngle = vtkMath::DegreesFromRadians(oldAngle);
 
 	double angle = newAngle - oldAngle;
-	cerr << "angle = " << angle << "\n";
+	//cerr << "angle = " << angle << "\n";
+	TRACE("angle = %g\n", angle);
 
 // COMMENT: {1/29/2004 9:41:02 PM}	//****  THIS NEEDS WORK
 // COMMENT: {1/29/2004 9:41:02 PM}	// SEEMS TO WORK COUNTERCLOCKWISE BUT NOT CLOCKWISE
@@ -466,19 +443,14 @@ void vtkPlaneWidget2::Spin(void)
 	this->Transform->Translate(-center[0], -center[1], -center[2]);
 
 	// Set the corners
-	vtkFloatingPointType oNew[3], pt1New[3], pt2New[3];
+	double oNew[3], pt1New[3], pt2New[3];
 	this->Transform->TransformPoint(o, oNew);
 	this->Transform->TransformPoint(pt1, pt1New);
 	this->Transform->TransformPoint(pt2, pt2New);
 
-	vtkFloatingPointType inv[3];
-	this->Transform->TransformPoint(this->m_InvisablePosX, inv);
-	for (int i=0; i < 3; ++i)
-	{
-		this->m_InvisablePosX[i] = inv[i];
-	}
+	this->Transform->TransformPoint(this->m_InvisablePosX, this->m_InvisablePosX);
 #if defined(_DEBUG)
-	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX);
+	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX[0], this->m_InvisablePosX[1], this->m_InvisablePosX[2]);
 #endif
 
 	this->PlaneSource->SetOrigin(oNew);
@@ -491,15 +463,15 @@ void vtkPlaneWidget2::Spin(void)
 
 double vtkPlaneWidget2::GetDeltaX(void)
 {
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-	vtkFloatingPointType *pt1 = this->PlaneSource->GetPoint1();
+	double *o = this->PlaneSource->GetOrigin();
+	double *pt1 = this->PlaneSource->GetPoint1();
 	double distance = sqrt(vtkMath::Distance2BetweenPoints(o, pt1));
 
 	// Unrotate in order to determine sign
 	this->Transform->Identity();
 	this->Transform->RotateZ(-this->GetAngle());
 
-	vtkFloatingPointType oNew[3], pt1New[3];
+	double oNew[3], pt1New[3];
 	this->Transform->TransformPoint(o, oNew);
 	this->Transform->TransformPoint(pt1, pt1New);
 	if (pt1New[0] > oNew[0])
@@ -518,28 +490,31 @@ double vtkPlaneWidget2::GetDeltaX(void)
 
 void vtkPlaneWidget2::SetDeltaX(double dx)
 {
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-	vtkFloatingPointType *pt1 = this->PlaneSource->GetPoint1();
-	vtkFloatingPointType *pt2 = this->PlaneSource->GetPoint2();
+	double *o = this->PlaneSource->GetOrigin();
+	double *pt1 = this->PlaneSource->GetPoint1();
+	double *pt2 = this->PlaneSource->GetPoint2();
 	double radians = atan2(pt1[1] - o[1], pt1[0] - o[0]);
 
 	//BUGBUG entering a value when length is negative
 	//causes a flip of entered sign
-	vtkFloatingPointType new_pt1[3];
+	double new_pt1[3];
 	new_pt1[0] = o[0] + (dx * cos(radians));
 	new_pt1[1] = o[1] + (dx * sin(radians));
 	new_pt1[2] = 0.0;
 
 	// check validity
-	vtkFloatingPointType v1[3], v2[3];
+	double v1[3], v2[3];
 	for (int i = 0; i < 3; ++i)
 	{
 		v1[i] = new_pt1[i] - o[i];
 		v2[i] = pt2[i] - o[i];
 	}
-	vtkFloatingPointType n[3];
+	double n[3];
 	vtkMath::Cross(v1, v2, n);
-	if ( vtkMath::Normalize(n) == 0.0 ) return;
+	if ( vtkMath::Normalize(n) == 0.0 )
+	{
+		return;
+	}
 
 	// if here ok to change
 	this->PlaneSource->SetPoint1(new_pt1);
@@ -549,8 +524,8 @@ void vtkPlaneWidget2::SetDeltaX(double dx)
 
 double vtkPlaneWidget2::GetDeltaY(void)
 {
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-	vtkFloatingPointType *pt2 = this->PlaneSource->GetPoint2();
+	double *o = this->PlaneSource->GetOrigin();
+	double *pt2 = this->PlaneSource->GetPoint2();
 	double distance = sqrt(vtkMath::Distance2BetweenPoints(o, pt2));
 
 	// Unrotate in order to determine sign
@@ -558,7 +533,7 @@ double vtkPlaneWidget2::GetDeltaY(void)
 	this->Transform->RotateZ(-this->GetAngle());
 
 	// Set the corners
-	vtkFloatingPointType oNew[3], pt2New[3];
+	double oNew[3], pt2New[3];
 	this->Transform->TransformPoint(o, oNew);
 	this->Transform->TransformPoint(pt2, pt2New);
 	if (pt2New[1] > oNew[1])
@@ -577,28 +552,29 @@ double vtkPlaneWidget2::GetDeltaY(void)
 
 void vtkPlaneWidget2::SetDeltaY(double dy)
 {
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-	vtkFloatingPointType *pt1 = this->PlaneSource->GetPoint1();
-	vtkFloatingPointType *pt2 = this->PlaneSource->GetPoint2();
+	double *o = this->PlaneSource->GetOrigin();
+	double *pt1 = this->PlaneSource->GetPoint1();
+	double *pt2 = this->PlaneSource->GetPoint2();
 	double radians = atan2(pt2[1] - o[1], pt2[0] - o[0]);
 
 	//BUGBUG entering a value when length is negative
 	//causes a flip of entered sign
-	vtkFloatingPointType new_pt2[3];
+	double new_pt2[3];
 	new_pt2[0] = o[0] + (dy * cos(radians));
 	new_pt2[1] = o[1] + (dy * sin(radians));
 	new_pt2[2] = 0.0;
 
 	// check validity
-	vtkFloatingPointType v1[3], v2[3];
+	double v1[3], v2[3];
 	for (int i = 0; i < 3; ++i)
 	{
 		v1[i] = pt1[i] - o[i];
 		v2[i] = new_pt2[i] - o[i];
 	}
-	vtkFloatingPointType n[3];
+	double n[3];
 	vtkMath::Cross(v1, v2, n);
-	if ( vtkMath::Normalize(n) == 0.0 ) {
+	if ( vtkMath::Normalize(n) == 0.0 )
+	{
 		return; // Bad plane coordinate system
 	}
 
@@ -609,41 +585,47 @@ void vtkPlaneWidget2::SetDeltaY(double dy)
 
 double vtkPlaneWidget2::GetRadians(void)
 {
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
+	double *origin = this->PlaneSource->GetOrigin();
+	//{{
+	double o[3];
+	o[0] = origin[0];
+	o[1] = origin[1];
+	o[2] = origin[2];
+	//}}
 	return atan2(this->m_InvisablePosX[1] - o[1], this->m_InvisablePosX[0] - o[0]);
 }
 
 void vtkPlaneWidget2::SetRadians(double radians)
 {
-	this->SetAngle(radians * vtkMath::DoubleRadiansToDegrees());
+	this->SetAngle(vtkMath::DegreesFromRadians(radians));
 }
 
 double vtkPlaneWidget2::GetAngle(void)
 {
-	return vtkMath::DoubleRadiansToDegrees() * this->GetRadians();
+	return vtkMath::DegreesFromRadians(this->GetRadians());
 }
 
 void vtkPlaneWidget2::SetAngle(double angle)
 {
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-	vtkFloatingPointType *pt1 = this->PlaneSource->GetPoint1();
-	vtkFloatingPointType *pt2 = this->PlaneSource->GetPoint2();
-	vtkFloatingPointType *center = this->PlaneSource->GetCenter();
+	double *o = this->PlaneSource->GetOrigin();
+	double *pt1 = this->PlaneSource->GetPoint1();
+	double *pt2 = this->PlaneSource->GetPoint2();
+	double *center = this->PlaneSource->GetCenter();
 
 	// Manipulate the transform to reflect the rotation
 	this->Transform->Identity();
-	this->Transform->Translate(center[0], center[1], center[2]);
+	this->Transform->Translate(o[0], o[1], o[2]);
 	this->Transform->RotateZ(angle - this->GetAngle());
-	this->Transform->Translate(-center[0], -center[1], -center[2]);
+	this->Transform->Translate(-o[0], -o[1], -o[2]);
 
 	// Set the corners
-	vtkFloatingPointType oNew[3], pt1New[3], pt2New[3];
+	double oNew[3], pt1New[3], pt2New[3];
 	this->Transform->TransformPoint(o, oNew);
 	this->Transform->TransformPoint(pt1, pt1New);
 	this->Transform->TransformPoint(pt2, pt2New);
 	this->Transform->TransformPoint(this->m_InvisablePosX, this->m_InvisablePosX);
 #if defined(_DEBUG)
-	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX);
+	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX[0], this->m_InvisablePosX[1], this->m_InvisablePosX[2]);
 #endif
 
 	this->PlaneSource->SetOrigin(oNew);
@@ -655,16 +637,16 @@ void vtkPlaneWidget2::SetAngle(double angle)
 	this->SizeHandles();
 }
 
-vtkFloatingPointType* vtkPlaneWidget2::GetModelOrigin(void)
+double* vtkPlaneWidget2::GetModelOrigin(void)
 {
 	return this->GetOrigin();
 }
 
 void vtkPlaneWidget2::SetModelOrigin(float x, float y)
 {
-	vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-	vtkFloatingPointType *pt1 = this->PlaneSource->GetPoint1();
-	vtkFloatingPointType *pt2 = this->PlaneSource->GetPoint2();
+	double *o = this->PlaneSource->GetOrigin();
+	double *pt1 = this->PlaneSource->GetPoint1();
+	double *pt2 = this->PlaneSource->GetPoint2();
 
 	//Manipulate the transform to reflect the rotation
 	this->Transform->Identity();
@@ -673,11 +655,11 @@ void vtkPlaneWidget2::SetModelOrigin(float x, float y)
 	this->m_InvisablePosX[0] += x - o[0];
 	this->m_InvisablePosX[1] += y - o[1];
 #if defined(_DEBUG)
-	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX);
+	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX[0], this->m_InvisablePosX[1], this->m_InvisablePosX[2]);
 #endif
 
 	//Set the corners
-	vtkFloatingPointType oNew[3], pt1New[3], pt2New[3];
+	double oNew[3], pt1New[3], pt2New[3];
 	this->Transform->TransformPoint(o,oNew);
 	this->Transform->TransformPoint(pt1,pt1New);
 	this->Transform->TransformPoint(pt2,pt2New);
@@ -701,7 +683,7 @@ void vtkPlaneWidget2::SizeHandles(void)
 		this->ValidPick = 1;
 	}
 
-	vtkFloatingPointType radius = this->vtk3DWidget::SizeHandles(.50);
+	double radius = this->vtk3DWidget::SizeHandles(.10);
 	for(int i = 0; i < 4; ++i)
 	{
 		this->HandleGeometry[i]->SetRadius(radius);
@@ -709,8 +691,8 @@ void vtkPlaneWidget2::SizeHandles(void)
 
 	if (this->CurrentRenderer)
 	{
-		vtkFloatingPointType *o = this->PlaneSource->GetOrigin();
-		vtkFloatingPointType scale = this->vtk3DWidget::SizeHandles(5.0);
+		double *o = this->PlaneSource->GetOrigin();
+		double scale = this->vtk3DWidget::SizeHandles(1.0);
 		if (this->XAxisActor)
 		{
 			this->XAxisActor->SetPosition(o);
@@ -722,27 +704,6 @@ void vtkPlaneWidget2::SizeHandles(void)
 			this->YAxisActor->SetPosition(o);
 			this->YAxisActor->SetScale(scale);
 			this->YAxisActor->SetOrientation(0, 0, this->GetAngle() + 90);
-		}
-		if (this->ActivePlaneSource)
-		{
-			vtkFloatingPointType *pt1 = this->PlaneSource->GetPoint1();
-			vtkFloatingPointType *pt2 = this->PlaneSource->GetPoint2();
-			double rad = this->GetRadians();
-
-			float npt1[3];
-			npt1[0] = scale*cos(rad) + o[0];
-			npt1[1] = scale*sin(rad) + o[1];
-			npt1[2] = 0.0;
-
-			float npt2[3];
-			npt2[0] = scale * cos(rad + vtkMath::Pi()/2) + o[0];
-			npt2[1] = scale * sin(rad + vtkMath::Pi()/2) + o[1];
-			npt2[2] = 0.0;
-
-			this->ActivePlaneSource->SetOrigin(o);
-			this->ActivePlaneSource->SetPoint1(npt1[0], npt1[1], npt1[2]);
-			this->ActivePlaneSource->SetPoint2(npt2[0], npt2[1], npt2[2]);
-			this->ActivePlaneSource->Update();
 		}
 #ifdef _DEBUG
 		if (this->m_VisHandleGeometry)
@@ -760,7 +721,7 @@ void vtkPlaneWidget2::PlaceWidget(void)
 	afxDump << "before vtkPlaneWidget2::PlaceWidget\n";
 	ostrstream oss;
 
-	this->PrintSelf(oss, 4);
+	this->PrintSelf(oss, vtkIndent(4));
 
 	oss << ends;
 	afxDump << oss.str() << "\n";
@@ -769,12 +730,27 @@ void vtkPlaneWidget2::PlaceWidget(void)
 
     this->Superclass::PlaceWidget();
 
-	vtkFloatingPointType* pt1 = this->GetPoint1();
+	//{{
+	double *o = this->GetOrigin();
+	this->GridOrigin[0] = o[0];
+	this->GridOrigin[1] = o[1];
+	this->GridOrigin[2] = o[2];
+
+	double *p1 = this->GetPoint1();
+	this->Deltas[0] = p1[0] - o[0];
+
+	double *p2 = this->GetPoint2();
+	this->Deltas[1] = p2[1] - o[1];
+
+	this->Deltas[2] = 0.;
+	//}}
+
+	double* pt1 = this->GetPoint1();
 	for (int i = 0 ; i < 3; ++i) {
 		this->m_InvisablePosX[i] = pt1[i];
 	}
 #if defined(_DEBUG)
-	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX);
+	this->m_VisHandleGeometry->SetCenter(this->m_InvisablePosX[0], this->m_InvisablePosX[1], this->m_InvisablePosX[2]);
 #endif
 }
